requirements.txt
nbval==0.10.0
jupyterlab
nbclient
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
fiona==1.9.5
fonttools==4.47.0
geopandas==0.14.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pipreqs==0.4.13
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.2
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9
pytest==7.3.2
pytest-mock==3.12.0
mock==4.0.3
